examples/agents/python/agno/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
agno==2.6.1
openai==2.8.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
jinja2==3.1.6
python-dotenv==1.0.1
PyJWT==2.10.1
zijus-tools==0.0.2
